# [celotool]Add fast mode to celotool invite