--- a/packages/celotool/src/cmds/account/invite.ts
+++ b/packages/celotool/src/cmds/account/invite.ts
@@ -12,24 +12,39 @@ export const describe = 'command for sending an invite code to a phone number'
 
 interface InviteArgv extends AccountArgv {
   phone: string
+  fast: boolean
 }
 
 export const builder = (yargs: Argv) => {
-  return yargs.option('phone', {
-    type: 'string',
-    description: 'Phone number to send invite code,',
-    demand: 'Please specify phone number to send invite code',
-  })
+  return yargs
+    .option('phone', {
+      type: 'string',
+      description: 'Phone number to send invite code,',
+      demand: 'Please specify phone number to send invite code',
+    })
+    .option('fast', {
+      type: 'boolean',
+      default: false,
+      description: "Don't download artifacts, use this for repeated invocations",
+      demand: 'Please specify phone number to send invite code',
+    })
 }
 
 export const handler = async (argv: InviteArgv) => {
-  await switchToClusterFromEnv(false)
   console.log(`Sending invitation code to ${argv.phone}`)
   const cb = async () => {
-    await execCmd(`yarn --cwd ../protocol run invite -n ${argv.celoEnv} -p ${argv.phone}`)
+    await execCmd(
+      `yarn --cwd ../protocol run invite -n ${argv.celoEnv} -p ${argv.phone} -f ${argv.fast}`
+    )
   }
   try {
-    await downloadArtifacts(argv.celoEnv)
+    if (argv.fast) {
+      console.log(`Fast mode is on, cluster won't be switched, artifacts won't be downloaded`)
+    } else {
+      console.log(`Fast mode is off, artifacts will be downloaded`)
+      await switchToClusterFromEnv(false)
+      await downloadArtifacts(argv.celoEnv)
+    }
     await portForwardAnd(argv.celoEnv, cb)
   } catch (error) {
     console.error(`Unable to send invitation code to ${argv.phone}`)