### Description

`celotool account invite` performs the following tasks

1. downloads the latest protocol artifacts
2. compiles the protocol scripts
3. changes Google cloud project
4. fetches the Twilio config
5. Sends the invite.

I added a new `fast` mode which assumes that step 1-4 have already been done in a previous invocation. This makes repeated invocations faster. This saves about 30 seconds. To avoid any confusions, the default path is still thorough and slow.

Also, in `invite.ts` script, the contracts were being fetched serially, fetching then in parallel saved about another 10 seconds. This is fixed for both fast and non-fast mode.

Overall, the thorough and slow invite time is down from ~80 seconds to ~70 seconds and repeated invite time is down from ~80 seconds to ~40 seconds.

### Tested

`time celotooljs account invite -e alfajoresstaging --phone +1650555555 --verbose`
`time celotooljs account invite -e alfajoresstaging --phone +16505555555 --verbose --fast`